+++ test/Unit/RuleSet/ExplicitRuleSetTestCase.php
@@ -13,19 +13,11 @@ declare(strict_types=1);
 
 namespace Ergebnis\PhpCsFixer\Config\Test\Unit\RuleSet;
 
-use Ergebnis\PhpCsFixer\Config\ExplicitRuleSet;
 use PhpCsFixer\Fixer;
 use PhpCsFixer\FixerConfiguration;
 
 abstract class ExplicitRuleSetTestCase extends AbstractRuleSetTestCase
 {
-    final public function testIsExplicitRuleSet(): void
-    {
-        $ruleSet = static::createRuleSet();
-
-        self::assertInstanceOf(ExplicitRuleSet::class, $ruleSet);
-    }
-
     final public function testRuleSetDoesNotConfigureRuleSets(): void
     {
         $ruleSet = static::createRuleSet();