+++ CHANGELOG.md
@@ -19,7 +19,6 @@ For a full diff see [`5.16.0...main`][5.16.0...main].
 
 - Allow implementations of `Config\RuleSet` to declare and configure custom fixers ([#872]), by [@localheinz]
 - Renamed `Config\RuleSet::targetPhpVersion()` to `Config\RuleSet::phpVersion()` ([#878]), by [@localheinz]
-- Renamed `Config\RuleSet\ExplicitRuleSet` to `Config\ExplicitRuleSet` ([#882]), by [@localheinz]
 
 ### Fixed
 
@@ -34,6 +33,7 @@ For a full diff see [`5.16.0...main`][5.16.0...main].
 - Removed `DOCTRINE_IGNORED_TAGS` constant from `Config\Ruleset\AbstractRuleSet` ([#875]), by [@localheinz]
 - Removed constructor from `Config\Ruleset\AbstractRuleSet` ([#876]), by [@localheinz]
 - Removed `Config\Ruleset\AbstractRuleSet` ([#877]), by [@localheinz]
+- Removed `Config\RuleSet\ExplicitRuleSet` ([#885]), by [@localheinz]
 
 ## [`5.16.0`][5.16.0]
 
@@ -1194,8 +1194,8 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#877]: https://github.com/ergebnis/php-cs-fixer-config/pull/877
 [#880]: https://github.com/ergebnis/php-cs-fixer-config/pull/880
 [#881]: https://github.com/ergebnis/php-cs-fixer-config/pull/881
-[#882]: https://github.com/ergebnis/php-cs-fixer-config/pull/882
 [#883]: https://github.com/ergebnis/php-cs-fixer-config/pull/883
+[#885]: https://github.com/ergebnis/php-cs-fixer-config/pull/885
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler